reporting_app/requirements.txt
Flask
gunicorn
boto3
psycopg2-binary